fix(models): keep tie policy within model package
